migrarte voxtral tts config/parser to transformers_utils
