+++ vllm_omni/__init__.py
@@ -28,6 +28,7 @@ except ModuleNotFoundError as exc:  # pragma: no cover - optional dependency
 
 # Register custom configs (AutoConfig, AutoTokenizer) as early as possible.
 from vllm_omni.transformers_utils import configs as _configs  # noqa: F401, E402
+from vllm_omni.transformers_utils import parsers as _parsers  # noqa: F401, E402
 
 from .config import OmniModelConfig
